Works, start refactoring
